// ui/file-tree/src/lib/models/git-file-status.enum.ts
export enum GitFileStatus {
  Unmodified = "UNMODIFIED",
  Modified = "MODIFIED",
  Staged = "STAGED",
  Untracked = "UNTRACKED",
  Conflicted = "CONFLICTED",
  Added = "ADDED",
  Deleted = "DELETED",
  Renamed = "RENAMED",
  Copied = "COPIED",
  Ignored = "IGNORED",
  Unknown = "UNKNOWN",
}